requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
oauthlib==3.2.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.3.1
tweepy==4.14.0
urllib3==2.2.1
